requirements.txt
Django>=1.3
djangorestframework>=2.3.8